requirements/base.txt
# Core Framework
Django>=5.1,<5.2
djangorestframework>=3.15,<4.0
djangorestframework-jsonapi>=7.0,<8.0

# Database
psycopg[binary]>=3.1,<4.0

# Filtering & Pagination
django-filter>=24.0,<25.0

# Background Jobs
celery>=5.4,<6.0
django-celery-beat>=2.6,<3.0
redis>=5.0,<6.0

# HTTP Client
httpx>=0.27,<1.0

# CORS
django-cors-headers>=4.4,<5.0

# Content Security Policy
django-csp>=4.0,<5.0

# Swagger / OpenAPI
drf-spectacular>=0.27,<1.0

# Error Tracking
sentry-sdk[django]>=2.0,<3.0

# Email
sendgrid>=6.11,<7.0

# AWS S3
boto3>=1.34,<2.0
django-storages>=1.14,<2.0

# Logging
python-json-logger>=2.0,<3.0

# Environment
python-dotenv>=1.0,<2.0

# Server
gunicorn>=22.0,<23.0

# Image Processing
Pillow>=10.0,<11.0

# Timezone
pytz>=2024.1
